requirements.txt
Django==4.2.1
requests==2.27.1
drf-spectacular==0.25.1
djangorestframework==3.14.0
pytest==7.3.1
pytest-django==4.5.2
pytest-mock==3.10.0
